requirements.txt
numpy
pandas
jsons
ta
# torch
pytorch-lightning
lightning
wandb
langchain
scikit-learn
joblib
scipy
tqdm
einops
tensorboard